requirements.txt
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.1.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0
django-allauth==0.38.0
django-bootstrap4==0.0.6
django-heroku==0.3.1
django-jet==1.0.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-taggit==1.3.0
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
oauth2client==4.1.3
oauthlib==2.1.0
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
pyuploadcare==2.5.0
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.23
whitenoise==3.3.1
